Split ServiceHandler to avoid linking to cors lib
